Move AutoFire settings page to Widget module settings. #154 [ci skip] - Change AutoFire settings page form to new FormRedux components form stack.
