+++ app/scripts/Widget/settings/index.js
@@ -1,0 +1,1 @@
+export { default as AutoFireForm } from './AutoFireForm.jsx'
