+++ app/scripts/pages/index.js
@@ -2,7 +2,6 @@ export { default as Login } from './Login/Login.jsx'
 export { default as Logout } from './Logout.jsx'
 export { default as ShowMobilization } from './ShowMobilization.jsx'
 export { default as NewBlock } from './NewBlock.jsx'
-export { default as AutoFireForm } from './AutoFireForm.jsx'
 export { default as DonationWidgetSettings } from './DonationWidgetSettings.jsx'
 export { default as RequireLogin } from './RequireLogin.js'
 export { default as CustomDomainWrapper } from './CustomDomainWrapper/CustomDomainWrapper.jsx'
